Moved Packager to own namespace for import reuse
